+++ crates/libs/lib-rpc/src/rpcs/group_rpc.rs
@@ -1,11 +1,14 @@
 use lib_core::ctx::Ctx;
 use lib_core::model::group::{Group, GroupBmc, GroupFilter, GroupForCreate, GroupForUpdate};
+use lib_core::model::schedule::{ScheduleBmc, ScheduleForCreate};
 use lib_core::model::ModelManager;
+use serde_json::json;
 
-use crate::{ParamsForCreate, ParamsForUpdate, ParamsIded, ParamsList};
-use crate::Result;
 use crate::router::RpcRouter;
 use crate::rpc_router;
+use crate::Error::UserNotAdmin;
+use crate::Result;
+use crate::{ParamsForCreate, ParamsForUpdate, ParamsIded, ParamsList};
 
 pub fn rpc_router() -> RpcRouter {
     rpc_router!(
@@ -15,6 +18,7 @@ pub fn rpc_router() -> RpcRouter {
         list_groups,
         update_group,
         delete_group,
+        check_group_exists,
     )
 }
 
@@ -23,15 +27,28 @@ pub async fn create_group(
     mm: ModelManager,
     params: ParamsForCreate<GroupForCreate>,
 ) -> Result<Group> {
+    if !&ctx.admin() {
+        return Err(UserNotAdmin);
+    }
     let ParamsForCreate { data } = params;
 
     let id = GroupBmc::create(&ctx, &mm, data).await?;
-    let group = GroupBmc::get(&ctx, &mm, id).await?;
+    let group: Group = GroupBmc::get(&ctx, &mm, id).await?;
+
+    let schedule = ScheduleForCreate {
+        user_id: None,
+        group_id: Some(id),
+        course: group.year,
+    };
+    ScheduleBmc::create(&ctx, &mm, schedule).await?;
 
     Ok(group)
 }
 
 pub async fn get_group(ctx: Ctx, mm: ModelManager, params: ParamsIded) -> Result<Group> {
+    if !&ctx.admin() {
+        return Err(UserNotAdmin);
+    }
     let ParamsIded { id } = params;
 
     let group = GroupBmc::get(&ctx, &mm, id).await?;
@@ -44,6 +61,9 @@ pub async fn list_groups(
     mm: ModelManager,
     params: ParamsList<GroupFilter>,
 ) -> Result<Vec<Group>> {
+    if !&ctx.admin() {
+        return Err(UserNotAdmin);
+    }
     let groups = GroupBmc::list(&ctx, &mm, params.filters, params.list_options).await?;
 
     Ok(groups)
@@ -54,6 +74,9 @@ pub async fn update_group(
     mm: ModelManager,
     params: ParamsForUpdate<GroupForUpdate>,
 ) -> Result<Group> {
+    if !&ctx.admin() {
+        return Err(UserNotAdmin);
+    }
     let ParamsForUpdate { id, data } = params;
 
     GroupBmc::update(&ctx, &mm, id, data).await?;
@@ -64,10 +87,36 @@ pub async fn update_group(
 }
 
 pub async fn delete_group(ctx: Ctx, mm: ModelManager, params: ParamsIded) -> Result<Group> {
+    if !&ctx.admin() {
+        return Err(UserNotAdmin);
+    }
     let ParamsIded { id } = params;
 
     let group = GroupBmc::get(&ctx, &mm, id).await?;
     GroupBmc::delete(&ctx, &mm, id).await?;
 
     Ok(group)
+}
+
+pub async fn check_group_exists(
+    ctx: Ctx,
+    mm: ModelManager,
+    params: ParamsForCreate<GroupForCreate>,
+) -> Result<bool> {
+    if !&ctx.admin() {
+        return Err(UserNotAdmin);
+    }
+    let ParamsForCreate { data } = params;
+
+    let filter_json = json!({
+        "course": {"$eq": data.course},
+        "stage": {"$eq": data.stage },
+        "year": {"$eq": data.year },
+        "letter": {"$eq": data.letter },
+    });
+    let filter = vec![serde_json::from_value(filter_json)?];
+
+    let groups = GroupBmc::list(&ctx, &mm, Some(filter), None).await?;
+
+    Ok(groups.len() > 0)
 }
